リファクタリング: キー割り当て画面からホットキー設定ダイアログを抜き出してコンポーネント化する
